atc: behaviour: public pipelines generate notifications
The only thing that's missing here is sending a `DELETE` to subscribers
in other teams when a public pipeline is later hidden. One way to solve
this is to also send the old value of `p.public` when a pipeline is
updated, and send DELETE if the user does not have access and it went
from public to not public.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>