db: add migrations for access_tokens
concourse/concourse#5851

Since ID tokens grow along with the claims of the user, we've run into
issues where users in hundreds of groups have ID tokens that, when sent
in a header, cause problems at load balancers that reject large headers.
Rather than using ID tokens as access tokens, Concourse can generate an
access token (random bytes) and store a mapping of that access token to
the claims on the ID token from dex. On each request, we'd query the
database to find the claims for that given access token.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>