atc/db: behaviour: archive pipelines on build finish
concourse/concourse#5835

There was some concern around the performance of the latest successful
build subquery.

We expect the most common use-case of automatic archiving to be when a
set_pipeline step is removed from a job. At the end of each build of a
job we now run a simpler query that does not need to hit the builds
table to figure out which pipelines it needs to archive. It simply uses
its current build ID as the "latest successful build".

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>