api: behaviour: cache user claims
concourse/concourse#5851

Based on some benchmarking, the performance of the API dropped fairly
substantially (at least for the /api/v1/user endpoint) - ~40% fewer
requests per second.

After caching was added, the benchmarks claim we're actually getting
better performance - this may just be that my CPU was a bit busier
before, but at least it doesn't seem to have a noticable performance
degradation.

The cache is an LRU cache with a 1MiB capacity for claims.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>