web: behaviour: favorite pipelines on dashboard
concourse/concourse#5439

Apologies for this long commit - lazy rendering sucks! We decided to
create a separate "computeLayout..." and "view..." function for the
favorite section since it differs a fair bit from the all pipelines
section groups. Headers are a bit more tricky than when there's just a
single team, since we need to position them absolutely (as there can be
multiple headers per row).

Teams are displayed "inline" with one another, so we display a header
above the first pipeline in each team. We also display a "team
(continued)" header if a team's pipelines carry over to the next row.

There are still a couple things to fix.

1. When you hover over a task in a favorited pipeline thumbnail, the
   corresponding pipeline in "all pipelines" will be styled the same way
   (i.e. as if it were hovered)
2. We don't display the "favorite pipelines" or "all pipelines" section
   headers yet.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Zoe Tian <ztian@pivotal.io>