web: behaviour: sticky build headers stack up
So when you scroll down into a particular substep of an across step,
you'll see the initial `across` header, the `vars` sub header, and the
substep header stuck to the top of the screen (but not overlapping).

I wanted to do this in Elm, but it just wasn't really feasible since
stacking dynamically sized elements with "position: sticky" requires
knowing the `offsetHeight`, and the Elm code would get a whole lot more
complicated trying hitting Browser.Dom.getElement for every header.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>