web: fix step header padding issue on firefox
On Firefox, step headers nested in a step-body appear with extra
whitespace (not sure why this didn't happen on Chrome, but oh well).

This is because c6f6a6116db8c695a0d9eec3bdfb0af6ef1c538b and
31a640598d3cf7e545c830bc060a84fdf3e846af introduced a CSS rule that
applies to all divs under a `step-body` (presumably under the assumption
that the only divs under step bodies are log lines - this was the case
up until we started nesting step headers in step bodies).

However, this rule is redundant, since
55a7712b2135d34cbb32d87aa1a50b977366232a introduced the same
`white-space: pre-wrap` on all `pre` elements - and log lines are nested
under a `pre` and so inherit from that rule. So, get rid of it!

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>