api: structure: move dependencies to AccessFactory
concourse/concourse#5851

As @jamieklassen pointed out in PR review, it was odd that the
AccessHandler had dependencies on a TokenVerifier and TeamsFetcher since
it just passed on the computed results to AccessFactory. While I'm
making changes here anyway, might as well make the dependencies a bit
more reasonable by having AccessFactory delegate to TokenVerifier and
TeamsFetcher directly. Now, AccessFactory is a bit less dumb than just
being a wrapper over the NewAccessor constructor.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>